Move seed check out of Account component
