+++ packages/example/src/App.tsx
@@ -3,6 +3,7 @@ import Account from "./components/Account";
 import Connect from "./components/Connect";
 import { getKeystoneStore, MobxStoreProvider } from "./mobx";
 import { trackInit } from "./tracking";
+import { AppCheck } from "./AppCheck";
 
 function App() {
   const mobxStore = getKeystoneStore();
@@ -13,6 +14,7 @@ function App() {
 
   return (
     <MobxStoreProvider value={mobxStore}>
+      <AppCheck/>
       <Connect/>
       <Account/>
     </MobxStoreProvider>
