+++ packages/example/src/components/Account/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { observer } from "mobx-react-lite";
 import { Loader, Modal } from 'semantic-ui-react';
 import Main from "./Main";
@@ -7,18 +7,10 @@ import { useBalance } from "../../hook/useBalance";
 import {useKeystoneStore} from "../../mobx";
 import { AccountBackground, AccountContainer, AccountLabel } from "./styles"
 import { AppStatus } from "../../mobx/runtime";
-import { useMFPCheck } from "../../hook/useMFPCheck";
 
 const Account = observer(() => {
-  const { runtime: {status}, resetStore } = useKeystoneStore();
+  const { runtime: {status} } = useKeystoneStore();
   const { balance, rate, refresh } = useBalance();
-  const isSameMFP = useMFPCheck();
-
-  useEffect(() => {
-    if(!isSameMFP){
-      resetStore();
-    }
-  }, [isSameMFP])
 
   return (
     <>
